Saving the register to allow return values switch
We now get back a return value from the switch to an enclave.
Having inline assembly + tyche playing around with registers messes up
the code.
I fixed it by pushing some of the registers on the stack and poping
them.
We need to figure out how to do that correctly