Making the example slightly more complicated.
We have a readonly region shared with the enclave as well.
Trying to modify it actually leads to a triple fault (not EPTViolation).

One way to make the example better would be to do a double enclave
transition.
First write a message, exit, print the message from the untrusted part,
transition back into the enclave, then attempt to write to the read-only
section and fault.